+++ packages/benchmark-js/src/BenchmarkGroup.ts
@@ -21,8 +21,13 @@ export class BenchmarkGroup {
         return this;
     }
 
-    public run() {
+    public run(): void {
         for (const instance of this.instances) instance.run();
         for (const instance of this.instances) instance.writeResult();
+    }
+
+    public writeTestersCode(): this {
+        for (const instance of this.instances) instance.writeTesterCode();
+        return this;
     }
 }